requirements.txt
# Core dependencies
crewai>=0.28.0
snowflake-connector-python>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# Supporting dependencies
langchain>=0.0.267
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0